requirements.txt
gradio
google-generativeai
python-dotenv

